Final Project finished
Created a cute website with my roommate who is a performing artist. Had him draw it out, and adjusted it under his discretion the entire time to bring it to a point of his satisfaction. Used only SASS to design, decided to try something creative with the navigation system and overall completed all requirements.